Moved error and success to utils folder
