+++ backend/PyMatcha/routes/api/auth.py
@@ -30,18 +30,18 @@ from itsdangerous import SignatureExpired
 from PyMatcha import ACCESS_TOKEN_EXPIRES
 from PyMatcha import redis
 from PyMatcha import REFRESH_TOKEN_EXPIRES
-from PyMatcha.errors import BadRequestError
-from PyMatcha.errors import ConflictError
-from PyMatcha.errors import NotFoundError
-from PyMatcha.errors import UnauthorizedError
-from PyMatcha.success import Success
-from PyMatcha.success import SuccessOutput
-from PyMatcha.success import SuccessOutputMessage
 from PyMatcha.utils import hash_password
 from PyMatcha.utils.confirm_token import confirm_token
 from PyMatcha.utils.confirm_token import generate_confirmation_token
 from PyMatcha.utils.decorators import validate_params
+from PyMatcha.utils.errors import BadRequestError
+from PyMatcha.utils.errors import ConflictError
+from PyMatcha.utils.errors import NotFoundError
+from PyMatcha.utils.errors import UnauthorizedError
 from PyMatcha.utils.mail import send_mail_text
+from PyMatcha.utils.success import Success
+from PyMatcha.utils.success import SuccessOutput
+from PyMatcha.utils.success import SuccessOutputMessage
 
 User = user.User
 get_user = user.get_user
