+++ backend/PyMatcha/routes/api/debug.py
@@ -7,13 +7,13 @@ from flask import current_app
 from flask import jsonify
 from flask import request
 from PyMatcha import redis
-from PyMatcha.errors import NotFoundError
 from PyMatcha.models.report import Report
 from PyMatcha.models.view import View
-from PyMatcha.success import Success
-from PyMatcha.success import SuccessDeleted
 from PyMatcha.utils.decorators import debug_token_required
 from PyMatcha.utils.decorators import validate_params
+from PyMatcha.utils.errors import NotFoundError
+from PyMatcha.utils.success import Success
+from PyMatcha.utils.success import SuccessDeleted
 
 debug_bp = Blueprint("debug", __name__)
