+++ backend/PyMatcha/utils/errors/__init__.py
@@ -16,9 +16,10 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
+from PyMatcha.utils.errors.badrequest import BadRequestError
+from PyMatcha.utils.errors.conflict import ConflictError
+from PyMatcha.utils.errors.forbidden import ForbiddenError
+from PyMatcha.utils.errors.notfound import NotFoundError
+from PyMatcha.utils.errors.unauthorized import UnauthorizedError
 
-from PyMatcha.success.created import SuccessCreated
-from PyMatcha.success.deleted import SuccessDeleted
-from PyMatcha.success.success import Success, SuccessOutput, SuccessOutputMessage
-
-__all__ = ["SuccessDeleted", "SuccessCreated", "Success", "SuccessOutput", "SuccessOutputMessage"]
+__all__ = ["BadRequestError", "ConflictError", "ForbiddenError", "NotFoundError", "UnauthorizedError"]
