+++ backend/PyMatcha/utils/errors/badrequest.py
@@ -16,11 +16,31 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
+from PyMatcha import application
+from PyMatcha.utils.errors.template import generate_error_json
 
-from PyMatcha.errors.badrequest import BadRequestError
-from PyMatcha.errors.conflict import ConflictError
-from PyMatcha.errors.forbidden import ForbiddenError
-from PyMatcha.errors.notfound import NotFoundError
-from PyMatcha.errors.unauthorized import UnauthorizedError
 
-__all__ = ["BadRequestError", "ConflictError", "ForbiddenError", "NotFoundError", "UnauthorizedError"]
+class BadRequestError(Exception):
+    """
+    This is the BadRequestError class for the Exception.
+    """
+
+    def __init__(self, msg: str, solution: str) -> None:
+        self.name = "Bad Request"
+        self.msg = msg
+        self.solution = solution
+        self.status_code = 400
+
+    pass
+
+
+@application.errorhandler(BadRequestError)  # type: ignore
+def generate_badrequest(error: BadRequestError) -> dict:
+    """
+    This is the 400 response creator. It will create a 400 response along with
+    a custom message and the 400 code
+
+    :param error: The error body
+    :return: Returns the response formatted
+    """
+    return generate_error_json(error, 400)
