+++ backend/PyMatcha/utils/errors/template.py
@@ -16,32 +16,30 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
+from typing import Any
 
-from PyMatcha import application
-from PyMatcha.errors.template import generate_error_json
+from flask import jsonify
 
 
-class BadRequestError(Exception):
+def generate_error_json(error: Any, status_code: int) -> dict:
     """
-    This is the BadRequestError class for the Exception.
+    This function is used to generate a json of the error passed
+
+    :param error: The error containing the message and solution
+    :param status_code: The status code of the error.
+    :return: Returns a json containing all the info
     """
-
-    def __init__(self, msg: str, solution: str) -> None:
-        self.name = "Bad Request"
-        self.msg = msg
-        self.solution = solution
-        self.status_code = 400
-
-    pass
-
-
-@application.errorhandler(BadRequestError)  # type: ignore
-def generate_badrequest(error: BadRequestError) -> dict:
-    """
-    This is the 400 response creator. It will create a 400 response along with
-    a custom message and the 400 code
-
-    :param error: The error body
-    :return: Returns the response formatted
-    """
-    return generate_error_json(error, 400)
+    success = False
+    json = {
+        "success": success,
+        "error": {
+            "type": error.__class__.__name__,
+            "name": error.name,
+            "message": error.msg,
+            "solution": error.solution,
+        },
+        "code": status_code,
+    }
+    resp = jsonify(json)
+    resp.status_code = status_code
+    return resp
