+++ backend/PyMatcha/utils/success/__init__.py
@@ -16,33 +16,10 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
+from PyMatcha.utils.success.created import SuccessCreated
+from PyMatcha.utils.success.deleted import SuccessDeleted
+from PyMatcha.utils.success.success import Success
+from PyMatcha.utils.success.success import SuccessOutput
+from PyMatcha.utils.success.success import SuccessOutputMessage
 
-from PyMatcha import application
-from PyMatcha.errors.template import generate_error_json
-
-
-class ConflictError(Exception):
-    """
-    This is the ConflictError class for the Exception.
-    """
-
-    def __init__(self, msg: str, solution: str) -> None:
-        self.name = "Conflict Error"
-        self.msg = msg
-        self.solution = solution
-        self.status_code = 409
-
-    pass
-
-
-@application.errorhandler(ConflictError)  # type: ignore
-def generate_conflict(error: ConflictError) -> dict:
-    """
-    This is the 409 response creator. It will create a 409 response along with
-    a custom message and the 409 code
-
-    :param error: The error body
-    :return: Returns the response formatted
-    """
-
-    return generate_error_json(error, 409)
+__all__ = ["SuccessDeleted", "SuccessCreated", "Success", "SuccessOutput", "SuccessOutputMessage"]
