+++ backend/PyMatcha/utils/success/created.py
@@ -16,20 +16,19 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
-
 from flask import jsonify
 
 
-def SuccessDeleted(message: str) -> dict:
+def SuccessCreated(message: str) -> dict:
     """
-    Will generate a SuccessDeleted message with the appropriate return code and jsonify it.
+    Will generate a SuccessCreated message with the appropriate return code and jsonify it.
 
     :param message: The message to include
 
     :return: Returns a json response
     """
 
-    status_code = 204
+    status_code = 201
     success = True
     json = {"success": success, "message": message, "code": status_code}
     resp = jsonify(json)
